// src/interfaces/themeInterfaces.ts
export interface DangerPaletteOptions {
  contrastText: string;
  dangerButton: string;
  dangerButtonHover: string;
}

export interface TertiaryPaletteOptions {
  main: string;
  seconed: string;
  primaryActiveHover: string;
  disableCommon: string;
}

export interface TransparentPaletteOptions {
  main: string;
  primaryActiveHover: string;
  disableCommon: string;
}

export interface PrimaryButtonPaletteOptions {
  main: string;
  seconed: string;
  primaryActiveHover: string;
  disableCommonTwo: string;
}

export interface SecondaryButtonPaletteOptions {
  main: string;
  seconed: string;
  primaryActiveHover: string;
  disableCommon: string;
}